Update download track logic
