+++ packages/web/src/services/track-download.ts
@@ -1,15 +1,29 @@
-import { CID, TrackDownload as TrackDownloadBase } from '@audius/common'
+import { TrackDownload as TrackDownloadBase } from '@audius/common'
 
 import { audiusBackendInstance } from './audius-backend/audius-backend-instance'
 
 class TrackDownload extends TrackDownloadBase {
   async downloadTrack(
-    cid: CID,
-    creatorNodeEndpoints: string[],
-    filename: string
+    { url, filename }: { url: string, filename: string }
   ) {
-    const audiusLibs = await this.audiusBackend.getAudiusLibs()
-    return audiusLibs.File.downloadCID(cid, creatorNodeEndpoints, filename)
+    // const audiusLibs = await this.audiusBackend.getAudiusLibs()
+    // return audiusLibs.File.downloadCID(cid, creatorNodeEndpoints, filename)
+    const response = await window.fetch(url)
+    if (!response.ok) {
+      throw new Error('Download unsuccessful')
+    }
+
+    const downloadURL = (url: string, filename: string) => {
+      if (document) {
+        const link = document.createElement('a')
+        link.href = url
+        link.target = '_blank'
+        link.download = filename
+        link.click()
+      }
+      throw new Error('No document found')
+    }
+    downloadURL(response.url, filename)
   }
 }
